pyproject.toml
[project]
name = "todolist-fastapi"
version = "0.1.0"
description = ""
authors = [
    { name = "Lucas Verri", email = "lfverri@sga.pucminas.br" }
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
]

[tool.poetry]
packages = [{ include = "app", from = "src" }]

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.2"
taskipy = "^1.14.1"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
httpx = "^0.28.1"

[tool.taskipy.tasks]
run = "python -m uvicorn src.app.app:app"
pre_test = "task lint"
test = "pytest --cov=app tests/ -vv"
post_test = "coverage html"
lint = "ruff check ."
format = "ruff format ."

[tool.pytest.ini_options]
pythonpath = "src"
addopts = "--cov=app -p no:warnings"

[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "PT"]

[tool.ruff.format]
preview = true
quote-style = "single"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
